build: :sparkles: add exports to index
